+++ spec/Suite/Plugin/MonkeySpec.php
@@ -44,7 +44,7 @@ describe("Monkey", function() {
     /**
      * Save current & reinitialize the Interceptor class.
      */
-    before(function() {
+    beforeAll(function() {
         $this->previous = Interceptor::instance();
         Interceptor::unpatch();
 
@@ -58,7 +58,7 @@ describe("Monkey", function() {
     /**
      * Restore Interceptor class.
      */
-    after(function() {
+    afterAll(function() {
         Interceptor::load($this->previous);
     });