+++ src/functions.php
@@ -6,12 +6,12 @@ use Kahlan\Suite;
 use Kahlan\Specification;
 use Kahlan\Allow;
 
-function before($closure) {
-    return Suite::current()->before($closure);
+function beforeAll($closure) {
+    return Suite::current()->beforeAll($closure);
 }
 
-function after($closure) {
-    return Suite::current()->after($closure);
+function afterAll($closure) {
+    return Suite::current()->afterAll($closure);
 }
 
 function beforeEach($closure) {